Always expect that the parameter in the route is not the entity type (ex: {developer_app}) in case of apps rather just {app}.
